requirements.txt
discord.py==2.6.3
python-dotenv